services/miner/requirements.txt
elasticsearch>=8.13.0,<9
PyYAML>=6.0
